Instance rendering working really well

+++ Ebony/src/systems/particleRenderer.cpp
@@ -19,12 +19,12 @@ namespace systems
 			}
 			
 			glBindBuffer(GL_ARRAY_BUFFER, particleEffect->particlePositionBuffer);
-			glBufferData(GL_ARRAY_BUFFER,	particleEffect->getMaxParticles() * 4 * sizeof(float), NULL, GL_STREAM_DRAW);
-			glBufferSubData(GL_ARRAY_BUFFER, 0, particleEffect->particleCount * 4 * sizeof(float), &particleEffect->particlePositionSizeData.at(0));
+			glBufferData(GL_ARRAY_BUFFER, static_cast<unsigned long long>(particleEffect->getMaxParticles()) * 4 * sizeof(float), NULL, GL_STREAM_DRAW);
+			glBufferSubData(GL_ARRAY_BUFFER, 0, static_cast<unsigned long long>(particleEffect->particleCount) * 4 * sizeof(float), &particleEffect->particlePositionSizeData.at(0));
 
 			glBindBuffer(GL_ARRAY_BUFFER, particleEffect->particleColorBuffer);
-			glBufferData(GL_ARRAY_BUFFER, particleEffect->getMaxParticles() * 4 * sizeof(float), NULL, GL_STREAM_DRAW);
-			glBufferSubData(GL_ARRAY_BUFFER, 0, particleEffect->particleCount * 4 * sizeof(float) , &particleEffect->particleColorData.at(0));
+			glBufferData(GL_ARRAY_BUFFER, static_cast<unsigned long long>(particleEffect->getMaxParticles()) * 4 * sizeof(float), NULL, GL_STREAM_DRAW);
+			glBufferSubData(GL_ARRAY_BUFFER, 0, static_cast<unsigned long long>(particleEffect->particleCount) * 4 * sizeof(float), &particleEffect->particleColorData.at(0));
 
 			graphics.DrawInstanced(s, particleEffect->texture, particleEffect->instancedVAO, particleEffect->particleCount);
 		}
